refactor: Controller, Service 분리 및 search 기능 GET으로 변경
