+++ src/main/java/umc/lightup/member/service/MemberCommandService.java
@@ -5,21 +5,10 @@ import umc.lightup.member.domain.Member;
 import umc.lightup.member.domain.Portfolio;
 import umc.lightup.member.dto.MemberRequestDTO;
 import umc.lightup.member.dto.MemberResponseDTO;
-import umc.lightup.member.enums.CredentialType;
 
 import java.util.List;
 
 public interface MemberCommandService {
-    Member joinMember(MemberRequestDTO.JoinDto request);
-    MemberResponseDTO.LoginResultDTO loginMember(MemberRequestDTO.PasswordLoginRequestDTO request);
-    MemberResponseDTO.LoginResultDTO loginMemberByGoogle(String authCode);
-    MemberResponseDTO.LoginResultDTO loginMemberByKakao(String authCode);
-    Member joinMemberByGoogle(String authCode);
-    Member joinMemberByKakao(String authCode);
-    Member addGoogleLogin(Member member, String authCode);
-    Member addKakaoLogin(Member member, String authCode);
-    Member addPasswordLogin(Member member, String password);
-    void removeCredential(Member member, CredentialType credentialType);
     Member getMember(String email);
     MemberResponseDTO.MemberInfoDTO getMember(long id, String viewerEmail);
     Member putMember(String email, MemberRequestDTO.ChangeDto request);
